+++ python/llm/src/bigdl/llm/vllm/entrypoints/openai/protocol.py
@@ -38,7 +38,7 @@ import time
 from typing import Dict, List, Literal, Optional, Union
 
 from pydantic import BaseModel, Field
-from bigdl.llm.vllm.utils.llm_utils import random_uuid
+from bigdl.llm.vllm.utils import random_uuid
 from bigdl.llm.vllm.entrypoints.openai.openai_protocol import (ErrorResponse, ModelPermission, ModelCard,
                                                      UsageInfo, LogProbs, ChatMessage, DeltaMessage)
