--- python/llm/src/bigdl/llm/vllm/utils/llm_utils.py
@@ -1,60 +1,0 @@
-#
-# Copyright 2016 The BigDL Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# Some parts of this file is adapted from
-# https://github.com/vllm-project/vllm/blob/main/vllm/utils.py
-# which is licensed under Apache License 2.0
-#
-# Copyright 2023 The vLLM team. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import uuid
-import socket
-from typing import List, Optional, Tuple, Union
-
-from transformers import AutoTokenizer, PreTrainedTokenizer, PreTrainedTokenizerFast
-
-from bigdl.llm.vllm.logger import init_logger
-from bigdl.llm.utils.common import invalidInputError
-
-logger = init_logger(__name__)
-
-
-class Counter:
-    def __init__(self, start: int = 0) -> None:
-        self.counter = start
-
-    def __next__(self) -> int:
-        i = self.counter
-        self.counter += 1
-        return i
-
-    def reset(self) -> None:
-        self.counter = 0
-
-
-def random_uuid() -> str:
-    return str(uuid.uuid4().hex)
